#150: Move QueryFilterFactory to its own file
- It used to be in the database helpers file, but this class is used across both backends
- The imports have been moved inside the staticmethod because the Flask TEST_BACKEND config option won't have been set by the time QueryFilterFactory has been imported, so the imports which decide which set of filters to import must be put in some code.
