+++ datagateway_api/common/database/helpers.py
@@ -202,42 +202,6 @@ class DeleteQuery(Query):
         self.commit_changes()
 
 
-class QueryFilterFactory(object):
-    @staticmethod
-    def get_query_filter(request_filter):
-        """
-        Given a filter return a matching Query filter object
-
-        This factory is not in common.filters so the created filter can be for the
-        correct backend. Moving the factory into that file would mean the filters would
-        be based off the abstract classes (because they're in the same file) which won't
-        enable filters to be unique to the backend
-
-        :param request_filter: dict - The filter to create the QueryFilter for
-        :return: The QueryFilter object created
-        """
-        filter_name = list(request_filter)[0].lower()
-        if filter_name == "where":
-            field = list(request_filter[filter_name].keys())[0]
-            operation = list(request_filter[filter_name][field].keys())[0]
-            value = request_filter[filter_name][field][operation]
-            return WhereFilter(field, value, operation)
-        elif filter_name == "order":
-            field = request_filter["order"].split(" ")[0]
-            direction = request_filter["order"].split(" ")[1]
-            return OrderFilter(field, direction)
-        elif filter_name == "skip":
-            return SkipFilter(request_filter["skip"])
-        elif filter_name == "limit":
-            return LimitFilter(request_filter["limit"])
-        elif filter_name == "include":
-            return IncludeFilter(request_filter["include"])
-        elif filter_name == "distinct":
-            return DistinctFieldFilter(request_filter["distinct"])
-        else:
-            raise FilterError(f" Bad filter: {request_filter}")
-
-
 def insert_row_into_table(table, row):
     """
     Insert the given row into its table
