+++ datagateway_api/common/helpers.py
@@ -6,7 +6,7 @@ from flask import request
 from flask_restful import reqparse
 from sqlalchemy.exc import IntegrityError
 
-from datagateway_api.common.database.helpers import QueryFilterFactory
+from datagateway_api.common.query_filter import QueryFilterFactory
 from datagateway_api.common.exceptions import (
     ApiError,
     AuthenticationError,
